+++ src/Pim/Bundle/EnrichBundle/Controller/MassEditActionController.php
@@ -2,6 +2,7 @@
 
 namespace Pim\Bundle\EnrichBundle\Controller;
 
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\Routing\RouterInterface;
@@ -61,6 +62,7 @@ class MassEditActionController extends AbstractDoctrineController
      * @param FormFactoryInterface       $formFactory
      * @param ValidatorInterface         $validator
      * @param TranslatorInterface        $translator
+     * @param EventDispatcherInterface   $eventDispatcher
      * @param ManagerRegistry            $doctrine
      * @param OperatorRegistry           $operatorRegistry
      * @param MassActionParametersParser $parametersParser
@@ -75,6 +77,7 @@ class MassEditActionController extends AbstractDoctrineController
         FormFactoryInterface $formFactory,
         ValidatorInterface $validator,
         TranslatorInterface $translator,
+        EventDispatcherInterface $eventDispatcher,
         ManagerRegistry $doctrine,
         OperatorRegistry $operatorRegistry,
         MassActionParametersParser $parametersParser,
@@ -89,6 +92,7 @@ class MassEditActionController extends AbstractDoctrineController
             $formFactory,
             $validator,
             $translator,
+            $eventDispatcher,
             $doctrine
         );